+++ packages/core/src/library/index.ts
@@ -1,3 +1,4 @@
+export * from './asset-keys'
 export * from './definitions'
 export * from './diff'
 export * from './hash'