+++ src/App.vue
@@ -7,6 +7,7 @@ import { TooltipProvider } from 'reka-ui'
 import { provideEditor, useI18n } from '@open-pencil/vue'
 import AppShell from '@/components/Shell/AppShell.vue'
 import AppToast from '@/components/Shell/AppToast.vue'
+import PublishLibraryDialog from '@/components/libraries/PublishLibraryDialog.vue'
 import RecoveryDialog from '@/components/recovery/RecoveryDialog.vue'
 import SettingsDialog from '@/components/settings/SettingsDialog.vue'
 import { useEditorStore } from '@/app/editor/active-store'
@@ -44,6 +45,7 @@ onMounted(() => {
     </AppShell>
     <SettingsDialog />
     <RecoveryDialog />
+    <PublishLibraryDialog />
     <AppToast />
   </TooltipProvider>
 </template>